change can only select one card at a time
Only one card will be selected at a time, making sure we can split up cards that are stacked.

TODO: either make cards displace other cards or have cards that are put on top of other cards render on top as well as being selected first when clicked. Overall I think we need to make a rework for the card functionality